pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tkintertools"
authors = [{ name = "Xiaokang2022", email = "2951256653@qq.com" }]
maintainers = [{ name = "Xiaokang2022", email = "2951256653@qq.com" }]
description = "tkintertools is a UI framework based on the Canvas class of tkinter"
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">=3.10"
keywords = [
    "tkinter",
    "tkintertools",
    "gui",
    "ui",
    "3d",
    "cross-platform",
    "themed",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: Chinese (Simplified)",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Widget Sets",
    "Typing :: Typed",
    "Development Status :: 4 - Beta",
]
dynamic = ["version"]

[project.optional-dependencies]
suggest = [
    "darkdetect==0.8.0; platform_system != 'Darwin'",
    "darkdetect[macos-listener]==0.8.0; platform_system == 'Darwin'",
    "pywinstyles==1.8; platform_system == 'Windows'",
    "pillow>=10.0.0",
]
full = [
    "darkdetect==0.8.0; platform_system != 'Darwin'",
    "darkdetect[macos-listener]==0.8.0; platform_system == 'Darwin'",
    "pywinstyles==1.8; platform_system == 'Windows'",
    "hPyT==1.3.1; platform_system == 'Windows'",
    "win32material==1.0.5; platform_system == 'Windows'",
    "matplotlib>=3.9.0",
    "pillow>=10.0.0",
]

[project.urls]
"Source code" = "https://github.com/Xiaokang2022/tkintertools"
Homepage = "https://xiaokang2022.github.io/tkintertools/"
Changelog = "https://xiaokang2022.github.io/tkintertools/CHANGELOG/"
"Release notes" = "https://xiaokang2022.github.io/tkintertools/news/3.0.0/News/"
"Bug tracker" = "https://github.com/Xiaokang2022/tkintertools/issues"

[tool.setuptools.dynamic]
version = { attr = "tkintertools.__version__" }

[tool.setuptools.packages.find]
include = ["tkintertools*"]
namespaces = false

# [tool.setuptools.package-data]
# pypackage = ["tkintertools/*/*.json"]

# python -m build
# python -m twine upload dist/*
